<tool.thriftcompiler> temporary solution : do not delete .svn directories
git-svn-id: file:///var/pocketcampus/trunk@1463 fd477b11-563a-408f-9d6e-eeb9ca009d0c

+++ tool/tool.thriftcompiler/src/org/pocketcampus/tool/thriftcompiler/FileUtils.java
@@ -87,6 +87,8 @@ public class FileUtils {
 	    if (dir.isDirectory()) {
 	        String[] children = dir.list();
 	        for (int i=0; i<children.length; i++) {
+	        	if(".svn".equals(children[i])) // do not delete .svn folders
+	        		continue;
 	            boolean success = deleteDir(new File(dir, children[i]));
 	            if (!success) {
 	                return false;
